Passage en double pour résoudre le probleme de stabilité numérique

+++ QuadricsBasedMeshSimplifier/quadric.cpp
@@ -14,7 +14,7 @@ Quadric::Quadric()
 		values_[val] = 0.f;
 }
 // ----------------------------------------------------------------------------
-Quadric::Quadric(float parValues[])
+Quadric::Quadric(double parValues[])
 {
 	for (size_t val = 0; val < MAX_VALUES; ++val)
 		values_[val] = parValues[val];
@@ -26,12 +26,10 @@ Quadric::~Quadric()
 // ----------------------------------------------------------------------------
 void Quadric::Add(const Quadric& parQuadric)
 {
-	//for (size_t val = 0; val < MAX_VALUES; ++val)
-//		values_[val] += parQuadric.values_[val];
 	Add(parQuadric.Values());
 }
 // ----------------------------------------------------------------------------
-void Quadric::Add(const float parValues[])
+void Quadric::Add(const double parValues[])
 {
 	for (size_t val = 0; val < MAX_VALUES; ++val)
 		values_[val] += parValues[val];
